pyproject.toml
[tool.poetry]
name = "von-neumann"
version = "0.1.0"
description = ""
authors = ["Stifled and Fickle"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.1"
pyjokes = "^0.6.0"
emoji = "^1.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
